Make the chat update when we send messages from other screens
